認証機能の移行の途中経過
